javaCore commit
