--- src/main/java/JavaCore/task2/Category.java
@@ -1,5 +1,0 @@
-package JavaCore.task2;
-
-public enum Category {
-    ELECTRONICS, CLOTHING, BOOKS, HOME, BEAUTY, TOYS
-}
