--- src/main/java/JavaCore/task2/Customer.java
@@ -1,18 +1,0 @@
-package JavaCore.task2;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class Customer {
-    private String customerId;
-    private String name;
-    private String email;
-    private LocalDateTime registeredAt;
-    private int age;
-    private String city;
-}
