--- src/main/java/JavaCore/task2/Order.java
@@ -1,18 +1,0 @@
-package JavaCore.task2;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.List;
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class Order  {
-    private String orderId;
-    private LocalDateTime orderDate;
-    private Customer customer;
-    private List<OrderItem> items;
-    private OrderStatus status;
-}
