--- src/main/java/JavaCore/task2/OrderItem.java
@@ -1,15 +1,0 @@
-package JavaCore.task2;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class OrderItem {
-    private String productName;
-    private int quantity;
-    private double price;
-    private Category category;
-}
