--- src/main/java/JavaCore/task2/OrderStatus.java
@@ -1,5 +1,0 @@
-package JavaCore.task2;
-
-public enum OrderStatus {
-    NEW, PROCESSING, SHIPPED, DELIVERED, CANCELLED
-}
