--- src/main/java/JavaCore/task3/Robot.java
@@ -1,5 +1,0 @@
-package JavaCore.task3;
-
-public enum Robot {
-    HEAD, TORSO, HAND,FEET
-}
