+++ src/main/java/javacore/task2/enums/Category.java
@@ -1,0 +1,5 @@
+package javacore.task2.enums;
+
+public enum Category {
+  ELECTRONICS, CLOTHING, BOOKS, HOME, BEAUTY, TOYS
+}
