+++ src/main/java/javacore/task2/enums/OrderStatus.java
@@ -1,0 +1,5 @@
+package javacore.task2.enums;
+
+public enum OrderStatus {
+  NEW, PROCESSING, SHIPPED, DELIVERED, CANCELLED
+}
