+++ src/main/java/javacore/task3/Robot.java
@@ -1,0 +1,5 @@
+package javacore.task3;
+
+public enum Robot {
+  HEAD, TORSO, HAND, FEET
+}
